moved commands
